\name{ausbeer}
\alias{ausbeer}
\docType{data}
\title{
Quarterly Australian Beer production
}
\description{
Total quarterly beer production in Australia (in megalitres) from 1956:Q1 to 2008:Q3.
}
\usage{data(ausbeer)}
\format{Quarterly time series of class \code{ts}.}

\source{
Australian Bureau of Statistics.
}
\examples{
data(ausbeer)
seasonplot(ausbeer)
}
\keyword{datasets}
