\name{fpp-package}
\alias{fpp-package}
\alias{fpp}
\docType{package}
\title{
Data sets for the workshop "Forecasting: principles and practice" by Rob J Hyndman}
\description{
All data sets for "Forecasting: principles and practice" workshop, Kandersteg, 20-22 June 2011.
}
\details{
\tabular{ll}{
Package: \tab fpp\cr
Type: \tab Package\cr
License: \tab GPL2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Rob J Hyndman

Maintainer: Rob.Hyndman@monash.edu
}
\keyword{ package }
\seealso{
\code{\link[forecast]{forecast}} package
}
